move apply code to a more logical place
