+++ staging/src/k8s.io/structured-merge-diff/applyv0/empty_element.go
