+++ staging/src/k8s.io/structured-merge-diff/applyv0/strategy/strategic_visitor.go
