+++ staging/src/k8s.io/structured-merge-diff/applyv0/type_element.go
